+++ src/components/slide/content/file-selector.jsx
@@ -82,7 +82,6 @@ function FileSelector({
       }
       return true;
     });
-
     onFilesChange({ target: { id: name, value: newFiles } });
   };